🗄️ Flatten pgmq queue migration
