+++ platform/scms/.env.sample
@@ -5,12 +5,9 @@ VITE_APP_CONFIG_DIRECTORY='.'
 VITE_APP_CONFIG_SCHEMA_DIRECTORY='../../'
 # Local dev only — Prisma CLI / migrations. Runtime DB URL comes from app-config (api.databaseUrl).
 DATABASE_URL=postgresql://journals:curvenote@localhost:5432/journals?statement_cache_size=0
-# Job queue provider. Default in local dev is now `supabase` (real pgmq + pg_net
-# against the Docker Postgres — see docker/postgres/Dockerfile). Set to `mock` to
-# use the in-memory queue instead (e.g. if your local container lacks pgmq).
-#QUEUES_PROVIDER=mock
-# Mock-provider only — push-to-drain loopback URL (default: http://localhost:${VITE_PORT}/v1/jobs/push-to-drain)
-#MOCK_QUEUE_DRAIN_URL=http://localhost:3031/v1/jobs/push-to-drain
-# Queue drain auth (push-to-drain Bearer token) lives in app-config: api.queueConsumerSecret
+# Job queue: Supabase pgmq (real pgmq + pg_net) against the Docker Postgres —
+# see docker/postgres/Dockerfile. The local image bundles pgmq/pg_net/pg_cron, so
+# `npm run db:rebuild` is required once. Queue drain auth (push-to-drain Bearer
+# token) lives in app-config: api.queueConsumerSecret.
 # Log every Prisma SQL query to the dev server console (ignored in production):
 #PRISMA_DEBUG_QUERIES=true
